refactor: rename component to just fields
